+++ ConditionalHelpers.hpp
@@ -18,28 +18,27 @@
 namespace onnx2trt
 {
 
-// Given a subgraph, find all of its external inputs (tensors entering the subgraph).
-// The result is returned in `subgraphInputs`, which is a map indexed by ITensor (a tensor entering the subgraph) and
-// with values indicating a set of external input indices.
-void getSubgraphInputs(std::vector<nvinfer1::ILayer*> const& newLayers,
-    std::unordered_map<nvinfer1::ITensor*, std::set<int32_t>>& subgraphInputs);
+using NodeName = std::string;
+using LayerName = std::string;
+using InputIndex = int32_t;
 
-// Given a subgraph, find all of its external outputs (tensors exiting the subgraph).
-// The result is returned in `subgraphInputs`, which is a map indexed by ITensor (a tensor exiting the subgraph) and
-// with values indicating a set of external outputs indices.
-void getSubgraphOutputs(const std::vector<nvinfer1::ILayer*>& newLayers,
-    std::unordered_map<nvinfer1::ITensor*, std::set<int32_t>>& subgraphOutputs,
-    const std::vector<std::string>& reportedOutputs);
+// A SubgraphPortsMap maps inputs' ports of each layer in an ONNX graph.
+using SubgraphPortsMap = std::unordered_map<const nvinfer1::ILayer*, std::unordered_set<InputIndex>>;
+
+// Given a subgraph, find all of its external inputs (tensors entering the subgraph).
+void getSubgraphInputs(const std::vector<nvinfer1::ILayer*>& newLayers, SubgraphPortsMap& externalInputs);
 
 // Take a snapshot of the network before and after parsing the subgraph and return a list
 // of newly added network layers.
 void importSubgraph(ImporterContext* ctx, ::ONNX_NAMESPACE::GraphProto const& subgraph,
     std::vector<nvinfer1::ILayer*>& newLayers, std::vector<TensorOrWeights>& subgraphTensors);
 
-using InputsMap = std::unordered_map<std::string, nvinfer1::IIfConditionalInputLayer*>;
+// An InputsMap tracks which IIfConditionalInputLayer we've added to a layer's inputs,
+// so that we can reuse them if needed.
+using InputsMap = std::unordered_map<LayerName, nvinfer1::IIfConditionalInputLayer*>;
 
 // Add IIfConditionalInputLayers to the inputs of the subgraph indicated by `subgraph`.
 void addIfInputLayers(ImporterContext* ctx, nvinfer1::IIfConditional* conditional, InputsMap& inputsMap,
-    const std::vector<nvinfer1::ILayer*>& newLayers);
+    const std::vector<nvinfer1::ILayer*>& newLayers, ::ONNX_NAMESPACE::NodeProto const* node);
 
 } // namespace onnx2trt